DisableRepoCommandHandler
